Added source generators for blueprint generation classes
