+++ src/BlueprintsNet.Generator.DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,6 +1,4 @@
-﻿using Light.GuardClauses;
-using BlueprintsNet.Generator.Services;
-using BlueprintsNet.Core.Models.Classes;
+﻿using BlueprintsNet.Generator.Services;
 using BlueprintsNet.Generator.Generators;
 
 namespace Microsoft.Extensions.DependencyInjection;
@@ -15,7 +13,8 @@ public static class ServiceCollectionExtensions
             .AddSingleton<IGenerator<Class>, ClassGenerator>()
             .AddSingleton<IGenerator<Constructor>, ConstructorGenerator>()
             .AddSingleton<IGenerator<Field>, FieldGenerator>()
-            .AddSingleton<IGenerator<Method>, MethodGenerator>();
+            .AddSingleton<IGenerator<Method>, MethodGenerator>()
+            .AddSingleton<IBlueprintGenerator, BlueprintGenerator>();
 
         return services;
     }
