+++ src/BlueprintsNet.Generator/Generators/BlueprintGenerators/ClassBlueprintGenerator.cs
@@ -1,5 +1,4 @@
-﻿using BlueprintsNet.Core.Models.Blueprints;
-
+﻿
 namespace BlueprintsNet.Generator.Generators;
 
 internal partial class BlueprintGenerator : BlueprintGeneratorBase
@@ -20,11 +19,6 @@ internal partial class BlueprintGenerator : BlueprintGeneratorBase
     }
 
     public override string Generate(BPMethodIn bp)
-    {
-        throw new NotImplementedException();
-    }
-
-    public override string Generate(BPLogicalAnd bp)
     {
         throw new NotImplementedException();
     }
