+++ src/BlueprintsNet.Generator/Generators/BlueprintGenerators/MathBlueprintGenerator.cs
@@ -1,6 +1,5 @@
-﻿using BlueprintsNet.Core.Models.Blueprints;
-
-namespace BlueprintsNet.Generator.Generators.BlueprintGenerators;
+﻿
+namespace BlueprintsNet.Generator.Generators;
 
 internal partial class BlueprintGenerator : BlueprintGeneratorBase
 {
