--- src/BlueprintsNet.SourceGenerators/BlueprintInterfaceGenerator.cs
@@ -1,87 +1,0 @@
-﻿using System;
-using System.Linq;
-using System.Text;
-using System.Collections.Generic;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Text;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-
-namespace BlueprintsNet.SourceGenerators
-{
-    [Generator]
-    public class BlueprintInterfaceGenerator : ISourceGenerator
-    {
-        public void Execute(GeneratorExecutionContext context)
-        {
-            var blueprints = ((BlueprintFinder)context.SyntaxReceiver).Blueprints;
-
-            var sourceBuilder = new StringBuilder();
-
-            sourceBuilder.Append(
-    @"using BlueprintsNet.Core.Models.Blueprints;
-
-namespace BlueprintsNet.Core
-{
-    public interface IBlueprintGenerator
-    {
-        string Generate(IBlueprint bp);
-");
-
-            blueprints.ForEach(blueprint => sourceBuilder.Append("        ")
-                                                .Append("string Generate(")
-                                                .Append(blueprint.Identifier.ValueText)
-                                                .AppendLine(" bp);")
-                                                .Append(Environment.NewLine));
-
-            sourceBuilder.Append(
-    @"    }
-}");
-
-            context.AddSource("IBlueprintGenerator.g.cs", SourceText.From(sourceBuilder.ToString(), Encoding.UTF8));
-        }
-
-        public void Initialize(GeneratorInitializationContext context)
-        {
-            context.RegisterForSyntaxNotifications(() => new BlueprintFinder());
-
-#if DEBUG
-            /*if (!Debugger.IsAttached)
-            {
-                Debugger.Launch();
-            }*/
-#endif
-        }
-
-        public class BlueprintFinder : ISyntaxReceiver
-        {
-            public List<ClassDeclarationSyntax> Blueprints { get; } = new();
-
-            public void OnVisitSyntaxNode(SyntaxNode syntaxNode)
-            {
-                if (syntaxNode is null)
-                {
-                    throw new ArgumentNullException(nameof(syntaxNode));
-                }
-
-                if (syntaxNode is ClassDeclarationSyntax declarationSyntax)
-                {
-                    var isAbstract = declarationSyntax.Modifiers
-                                         .FirstOrDefault(token => token.ValueText == "abstract")
-                        != default;
-                    var isBlueprint = declarationSyntax.BaseList?.Types
-                                        .FirstOrDefault(type =>
-                                        {
-                                            var identifierText = ((IdentifierNameSyntax)type.Type)?.Identifier.ValueText;
-
-                                            return identifierText == "IBlueprint" || identifierText.StartsWith("BP");
-                                        })
-                        != null;
-                    if (isBlueprint && !isAbstract)
-                    {
-                        Blueprints.Add(declarationSyntax);
-                    }
-                }
-            }
-        }
-    }
-}
