+++ src/BlueprintsNet.SourceGenerators/Generators/BlueprintInterfaceGenerator.cs
@@ -1,0 +1,52 @@
+﻿using System;
+using System.Text;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Text;
+
+namespace BlueprintsNet.SourceGenerators
+{
+    [Generator]
+    public class BlueprintInterfaceGenerator : ISourceGenerator
+    {
+        public void Execute(GeneratorExecutionContext context)
+        {
+            var blueprints = ((BlueprintFinder)context.SyntaxReceiver).Blueprints;
+
+            var sourceBuilder = new StringBuilder();
+
+            sourceBuilder.Append(
+@"using BlueprintsNet.Core.Models.Blueprints;
+
+namespace BlueprintsNet.Generator.Generators
+{
+    public interface IBlueprintGenerator
+    {
+        string Generate(IBlueprint bp);
+");
+
+            blueprints.ForEach(blueprint => sourceBuilder.Append("        ")
+                                                .Append("string Generate(")
+                                                .Append(blueprint.Identifier.ValueText)
+                                                .AppendLine(" bp);")
+                                                .Append(Environment.NewLine));
+
+            sourceBuilder.Append(
+    @"    }
+}");
+
+            context.AddSource("IBlueprintGenerator.g.cs", SourceText.From(sourceBuilder.ToString(), Encoding.UTF8));
+        }
+
+        public void Initialize(GeneratorInitializationContext context)
+        {
+            context.RegisterForSyntaxNotifications(() => new BlueprintFinder());
+
+#if DEBUG
+            /*if (!Debugger.IsAttached)
+            {
+                Debugger.Launch();
+            }*/
+#endif
+        }
+    }
+}
