+++ tests/BlueprintsNet.Generator.Tests/Generators/ConstructorGeneratorTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using BlueprintsNet.Core;
 using BlueprintsNet.Generator.Generators;
 
 namespace BlueprintsNet.Generator.Tests.Generators;
@@ -7,11 +8,14 @@ namespace BlueprintsNet.Generator.Tests.Generators;
 [TestFixture(TestOf = typeof(ConstructorGenerator))]
 public class ConstructorGeneratorTests
 {
+    private IBlueprintGenerator _bpGenerator;
     private ConstructorGenerator _subject;
 
     [SetUp]
     public void SetUp()
     {
-        _subject = new ConstructorGenerator();
+        _bpGenerator = A.Fake<IBlueprintGenerator>(options => options.Strict());
+
+        _subject = new ConstructorGenerator(_bpGenerator);
     }
 }
